barkive_flask/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
PyMySQL==1.1.1
PyYAML==6.0.2
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
